+++ .github/readme/synth.py
@@ -4,7 +4,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#      http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,19 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# Format: //devtools/kokoro/config/proto/build.proto
+"""This script is used to synthesize generated the README for this library."""
 
-# Download secrets from Cloud Storage.
-gfile_resources: "/bigstore/cloud-devrel-kokoro-resources/java-docs-samples"
+from synthtool.languages import java
 
-# Configure the docker image for kokoro-trampoline.
-env_vars: {
-    key: "TRAMPOLINE_IMAGE"
-    value: "gcr.io/cloud-devrel-kokoro-resources/java8"
-}
-
-# Tell trampoline which tests to run.
-env_vars: {
-    key: "TRAMPOLINE_BUILD_FILE"
-    value: "github/java-pubsub/.kokoro/run_samples_tests.sh"
-}
+java.custom_templates(["java_library/README.md"])